# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Mynaani API
  description: Geragogy-compliant learning platform API
  version: 1.0.0
  contact:
    name: Mynaani Engineering
servers:
  - url: https://noni-api-production.up.railway.app/api/v1
    description: Production
paths:
...
  /ui-envelope/{state_id}:
    get:
      summary: Get UI state envelope
      parameters:
        - name: state_id
          in: path
          required: true
          schema: { type: string, maxLength: 64, pattern: "^[a-zA-Z0-9_.-]+$" }
      responses:
        "200":
          description: Envelope found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UIStateEnvelope"
        "404":
          description: State undefined
...
  /curriculum/{module}/{unit_id}:
    get:
      summary: Curriculum unit content
      parameters:
        - name: module
          in: path
          required: true
          schema: { type: integer }
        - name: unit_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Unit content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResponse"
...
  /auth/session:
    get:
      summary: Session resolver
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Session active
        "401":
          description: No valid session
...
  /me:
    get:
      summary: Account settings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Account data
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UIStateEnvelope:
      type: object
      properties:
        state_id: { type: string }
        authorized_components: { type: array, items: { type: string } }
        interaction_limits: { type: object }
        layout_constraints: { type: object }
        transition_permissions: { type: array }
...
    LessonResponse:
      type: object
      properties:
        unit_id: { type: string }
        module: { type: integer }
        pages: { type: array }